+++ lib/elixir/lib/access.ex
@@ -1,13 +1,13 @@
-import Kernel, except: [access: 2]
-
 defprotocol Access do
   @moduledoc """
-  The Access protocol is the underlying protocol invoked
-  when the brackets syntax is used. For instance, `foo[bar]`
-  is translated to `access foo, bar` which, by default,
-  invokes the `Access.access` protocol.
+  The Access protocol is used by `foo[bar]` and also
+  empowers the nested update functions in Kernel.
 
-  This protocol is implemented by default for Lists, Maps
+  For instance, `foo[bar]` translates `Access.get(foo, bar)`.
+  `Kernel.get_in/2`, `Kernel.put_in/3` and `Kernel.update_in/3`
+  are also all powered by the Access protocol.
+
+  This protocol is implemented by default for keywords, maps
   and dictionary like types:
 
       iex> keywords = [a: 1, b: 2]
@@ -23,14 +23,19 @@ defprotocol Access do
       "★☆"
 
   The key access must be implemented using the `===` operator.
-  This protocol is limited and is implemented only for the
-  following built-in types: keywords, maps and functions.
   """
 
   @doc """
-  Receives the element being accessed and the access item.
+  Accesses the given key in the container.
   """
   def get(container, key)
+
+  @doc """
+  Updates the given key in container with a function.
+
+  In case the key is not set, invokes the function passing nil.
+  """
+  def update(container, key, fun)
 
   @doc false
   Kernel.def access(container, key) do
@@ -47,7 +52,20 @@ defimpl Access, for: List do
   end
 
   def get(_dict, key) do
-    raise ArgumentError, "the access protocol for lists expect the key to be an atom, got: #{inspect key}"
+    raise ArgumentError,
+      "the access protocol for lists expect the key to be an atom, got: #{inspect key}"
+  end
+
+  def update([{key, v}|t], key, fun) when is_atom(key) do
+    [{key, fun.(v)}|t]
+  end
+
+  def update([h|t], key, fun) do
+    [h|update(t, key, fun)]
+  end
+
+  def update([], key, fun) when is_atom(key) do
+    [{key, fun.(nil)}]
   end
 end
 
@@ -58,6 +76,16 @@ defimpl Access, for: Map do
       :error -> nil
     end
   end
+
+  def update(map, key, fun) do
+    value =
+      case :maps.find(key, map) do
+        {:ok, value} -> value
+        :error -> nil
+      end
+
+    :maps.put(key, fun.(value), map)
+  end
 end
 
 defimpl Access, for: Atom do
@@ -66,7 +94,21 @@ defimpl Access, for: Atom do
   end
 
   def get(atom, _) do
-    raise "The access protocol can only be invoked for atoms at " <>
-      "compilation time, tried to invoke it for #{inspect atom}"
+    undefined(atom)
+  end
+
+  def update(nil, _, fun) do
+    fun.(nil)
+  end
+
+  def update(atom, _key, _fun) do
+    undefined(atom)
+  end
+
+  defp undefined(atom) do
+    raise Protocol.UndefinedError,
+      protocol: @protocol,
+      value: atom,
+      description: "only the nil atom is supported"
   end
 end