+++ CHANGELOG.md
@@ -22,7 +22,7 @@
   * [Kernel] `set_elem/3` is deprecated in favor of `put_elem/3`
 
 * Backwards incompatible changes
-  * [Access] `Kernel.access/2` no longer exists and the `Access` protocol now uses `get/2` instead of `access/2`
+  * [Access] `Kernel.access/2` no longer exists and the `Access` protocol now requires `get/2` (instead of `access/2`) and `update/3` to be implemented
   * [Kernel] Retrieving docs as `module.__info__(:docs)` is deprecated, instead `Code.get_docs/2` must be used
   * [Mix] `mix new` no longer generates a supevision tree by default, please pass `--sup` instead